requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.1
geventhttpclient==1.4.5
greenlet==1.0.0
idna==2.9
inflection==0.4.0
itsdangerous==1.1.0
itypes==1.1.0
jdatetime==3.6.2
Jinja2==2.11.1
locust==1.4.1
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==1.4.6
numpy==1.18.2
openapi-codec==1.3.2
packaging==20.3
pandas==0.24.1
psutil==5.8.0
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==21.0.1
querystring-parser==1.2.4
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
xlrd==1.2.0
zope.event==4.5.0
zope.interface==5.2.0
